refactor: refact deleteRealEmoji logic
